refactor(frontend): extract lecture watch page utils
